ComponentModel.Search -> Search.ComponentModel
