+++ Kinetix/Kinetix.ComponentModel/Kinetix.ComponentModel.csproj
@@ -148,17 +148,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReferenceAttribute.cs" />
     <Compile Include="ReferencedTypeAttribute.shared.cs" />
-    <Compile Include="Search\Criteria.cs" />
-    <Compile Include="Search\FacetItem.cs" />
-    <Compile Include="Search\FacetListInput.cs" />
-    <Compile Include="Search\FacetOutput.cs" />
-    <Compile Include="Search\FacetSelection.cs" />
-    <Compile Include="Search\GroupResult.cs" />
-    <Compile Include="Search\GroupResult{TDocument}.cs" />
-    <Compile Include="Search\QueryInput.cs" />
-    <Compile Include="Search\QueryInput{TCriteria}.cs" />
-    <Compile Include="Search\QueryOutput{TDocument}.cs" />
-    <Compile Include="Search\QueryOutput.cs" />
     <Compile Include="SR.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
