+++ Kinetix/Kinetix.Search/ComponentModel/FacetItem.cs
@@ -1,0 +1,32 @@
+﻿namespace Kinetix.Search.ComponentModel {
+
+    /// <summary>
+    /// Item de facette.
+    /// </summary>
+    public struct FacetItem {
+
+        /// <summary>
+        /// Code de la facette.
+        /// </summary>
+        public string Code {
+            get;
+            set;
+        }
+
+        /// <summary>
+        /// Libellé de l'item.
+        /// </summary>
+        public string Label {
+            get;
+            set;
+        }
+
+        /// <summary>
+        /// Nombre d'éléments pour l'item.
+        /// </summary>
+        public long Count {
+            get;
+            set;
+        }
+    }
+}
