+++ Kinetix/Kinetix.Search/ComponentModel/QueryOutput.cs
@@ -1,25 +1,16 @@
 ﻿using System.Collections.Generic;
 
-namespace Kinetix.ComponentModel.Search {
+namespace Kinetix.Search.ComponentModel {
 
     /// <summary>
     /// Sortie d'une recherche avancée.
     /// </summary>
-    /// <typeparam name="TDocument">Le type du document.</typeparam>
-    public class QueryOutput<TDocument> {
-
-        /// <summary>
-        /// Liste de résultats (cas d'une recherche sans groupe).
-        /// </summary>
-        public ICollection<TDocument> List {
-            get;
-            set;
-        }
+    public class QueryOutput {
 
         /// <summary>
         /// Groupe de liste de résultats (cas d'une recherche avec groupe).
         /// </summary>
-        public ICollection<GroupResult<TDocument>> Groups {
+        public ICollection<GroupResult> Groups {
             get;
             set;
         }
