+++ Kinetix/Kinetix.Search/Elastic/ElasticStore.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Kinetix.ComponentModel.Search;
 using Kinetix.Search.ComponentModel;
 using Kinetix.Search.Contract;
 using Kinetix.Search.Elastic.Faceting;
