+++ Kinetix/Tests/Kinetix.Search.Test/SearchBrokerTest/AdvancedCountTest.cs
@@ -1,5 +1,5 @@
-﻿using Kinetix.ComponentModel.Search;
-using Kinetix.Search.Broker;
+﻿using Kinetix.Search.Broker;
+using Kinetix.Search.ComponentModel;
 using Kinetix.Search.Model;
 using Kinetix.Search.Test.Dum;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
