+++ Kinetix/Tests/Kinetix.Search.Test/SearchBrokerTest/PortfolioTest.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq;
-using Kinetix.ComponentModel.Search;
 using Kinetix.Search.Broker;
+using Kinetix.Search.ComponentModel;
 using Kinetix.Search.Model;
 using Kinetix.Search.Test.Dum;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
